新增代码

+++ hj_service_api/src/main/java/com/hj/bo/article/ArticleCategoryBo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ArticleCategoryBo {
 
-    private Integer articleCategoryId;
+    private Long articleCategoryId;
 
     /**
      * 文章分类标题
