+++ hj_service_api/src/main/java/com/hj/bo/user/FeedbackBo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class FeedbackBo {
 
-    private Integer feedbackId;
+    private Long feedbackId;
 
     private String reply;
 }
