+++ hj_service_api/src/main/java/com/hj/bo/user/MemberInterestsBo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class MemberInterestsBo {
 
-    private Integer interestsId;
+    private Long interestsId;
 
     private String name;
