+++ hj_service_api/src/main/java/com/hj/bo/user/UserLabelBo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UserLabelBo {
 
-    private Integer labelId;
+    private Long labelId;
 
     private String labelName;
