+++ hj_service_api/src/main/java/com/hj/vo/article/ArticleCategoryVo.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ArticleCategoryVo {
 
-    private Integer articleCategoryId;
+    private Long articleCategoryId;
 
     /**
      * 文章分类标题
