+++ hj_service_api/src/main/java/com/hj/vo/user/FeedbackCategory1Vo.java
@@ -8,9 +8,9 @@ import java.util.List;
 @Data
 public class FeedbackCategory1Vo {
 
-    private Integer feedbackCategoryId;
+    private Long feedbackCategoryId;
 
-    private Integer pid;
+    private Long pid;
 
     private String cateName;
 
@@ -24,7 +24,7 @@ public class FeedbackCategory1Vo {
 
     private Integer level;
 
-    private Integer merId;
+    private Long merId;
 
     private Date createTime;
