+++ hj_system/src/main/java/com/hj/service/user/MemberInterestsService.java
@@ -40,5 +40,5 @@ public interface MemberInterestsService extends IService<MemberInterests> {
      *
      * @param interestsId
      */
-    void deleteMemberInterests(Integer interestsId);
+    void deleteMemberInterests(Long interestsId);
 }
